Removed checks for keys

+++ inc/keycard.h
@@ -17,7 +17,7 @@
 
 void setup_keycard(void);
 void read_key(void);
-void key_check(uint32_t key_bit);
+void key_check(uint8_t key_bit);
 void update_key_state(void);
